rename files like in JS SDK
