+++ src/__tests__/browserSuites/events.spec.js
