+++ src/__tests__/browserSuites/manager.spec.js
