+++ src/__tests__/browserSuites/shared-instantiation.spec.js
