+++ src/__tests__/browserSuites/telemetry.spec.js
