+++ src/__tests__/browserSuites/use-beacon-api.debug.spec.js
