+++ src/__tests__/browserSuites/use-beacon-api.spec.js
