+++ src/__tests__/browserSuites/user-consent.spec.js
