+++ src/__tests__/online/browser.spec.js
@@ -1,22 +1,22 @@
 import tape from 'tape-catch';
 import fetchMock from '../testUtils/fetchMock';
 import { url } from '../testUtils';
-import evaluationsSuite from './evaluations.spec';
-import impressionsSuite from './impressions.spec';
-import impressionsSuiteDebug from './impressions.debug.spec';
-import telemetrySuite from './telemetry.spec';
-import impressionsListenerSuite from './impressions-listener.spec';
-import readinessSuite from './readiness.spec';
-import readyFromCache from './ready-from-cache.spec';
-import { withoutBindingTT /*, bindingTT */ } from './events.spec';
-import sharedInstantiationSuite from './shared-instantiation.spec';
-import managerSuite from './manager.spec';
-import ignoreIpAddressesSettingSuite from './ignore-ip-addresses-setting.spec';
-import useBeaconApiSuite from './use-beacon-api.spec';
-import useBeaconDebugApiSuite from './use-beacon-api.debug.spec';
-import readyPromiseSuite from './ready-promise.spec';
-import fetchSpecificSplits from './fetch-specific-splits.spec';
-import userConsent from './user-consent.spec';
+import evaluationsSuite from '../browserSuites/evaluations.spec';
+import impressionsSuite from '../browserSuites/impressions.spec';
+import impressionsSuiteDebug from '../browserSuites/impressions.debug.spec';
+import telemetrySuite from '../browserSuites/telemetry.spec';
+import impressionsListenerSuite from '../browserSuites/impressions-listener.spec';
+import readinessSuite from '../browserSuites/readiness.spec';
+import readyFromCache from '../browserSuites/ready-from-cache.spec';
+import { withoutBindingTT /*, bindingTT */ } from '../browserSuites/events.spec';
+import sharedInstantiationSuite from '../browserSuites/shared-instantiation.spec';
+import managerSuite from '../browserSuites/manager.spec';
+import ignoreIpAddressesSettingSuite from '../browserSuites/ignore-ip-addresses-setting.spec';
+import useBeaconApiSuite from '../browserSuites/use-beacon-api.spec';
+import useBeaconDebugApiSuite from '../browserSuites/use-beacon-api.debug.spec';
+import readyPromiseSuite from '../browserSuites/ready-promise.spec';
+import fetchSpecificSplits from '../browserSuites/fetch-specific-splits.spec';
+import userConsent from '../browserSuites/user-consent.spec';
 
 import { settingsValidator } from '../../settings';
 import splitChangesMock1 from '../mocks/splitchanges.since.-1.json';
